Pass auto_unbox as a variable to json function
Currently the `res$json` function automatically unboxes the object you're converting to JSON. 

In my use case I wanted to send an array with one value. But a list with one value is "unboxed" and just returns the one value.

In the end I manually converted to json with `auto_unbox = FALSE` and used `res$body` to send the response.  This can all be avoided if we can pass a variable for `auto_unbox`.